fix: SAML authn request fails due to cert signature validation
- UAA integration test failure:
```
SimpleSAML\Error\Error: UNHANDLEDEXCEPTION
...Caused by: SimpleSAML\Error\Exception: Missing certificate in metadata for '***.cloudfoundry-saml-login'
```
- according to simplesamlphph v1->v2 migration guide (https://simplesamlphp.org/docs/2.0/simplesamlphp-upgrade-notes-2.0.html):
```
SAML AuthnRequests that are signed will have their signature validated unless specifically disabled by setting validate.authnrequest to false . If unset (or set to true ) signatures will be validated if present and requests not passing validation will be refused.
```
this validation is now on by default if authn request contains signature (which implies that this is off by default in v1).
- Hence, turn off signature validation in v2 to restore to the v1 behavior for now.

[#185279615]

Co-authored-by: Danny Faught <dfaught@vmware.com>