fix: SAML single logout request fails due to cert signature validation
- UAA integration test failure:
```
SimpleSAML\Error\Error: UNHANDLEDEXCEPTION
Caused by: SimpleSAML\Error\Exception: Missing certificate in metadata for 'cloudfoundry-saml-login'
```
- The fix is similar to 81949c2
where the authn request signature validation is turned off.
- However, unlike "validate.authnrequest" whose change of default value from false to true
is explicitly documented in the simplesamlphph v1->v2 migration guide (https://simplesamlphp.org/docs/2.0/simplesamlphp-upgrade-notes-2.0.html),
"validate.logout"'s default value change is not documented in the migration guide. But we think it is highly
likely that its default has changed from false to true as well. So in this commit, we turn it off
to maintain what we assume to be the status quo.
- Further, it is difficult for the deployed simple saml server to obtain the SP metadata from localhost
(served by the locally started UAA when we run integration tests locally) needed to perform this validation.

[#185279615]

Co-authored-by: Hongchol Sinn <hsinn@vmware.com>